Changed Pointer to use usize instead of u64.

+++ src/memory_reader/memory_map_region.rs
@@ -28,7 +28,7 @@ impl MemoryMapRegion {
         }
     }
 
-    pub fn size_bytes(&self) -> u64 {
+    pub fn size_bytes(&self) -> usize {
         self.end - self.start
     }
